+++ src/Symfony/Component/Notifier/Bridge/AmazonSns/Tests/AmazonSnsTransportTest.php
@@ -15,6 +15,7 @@ use AsyncAws\Sns\Result\PublishResponse;
 use AsyncAws\Sns\SnsClient;
 use Symfony\Component\Notifier\Bridge\AmazonSns\AmazonSnsOptions;
 use Symfony\Component\Notifier\Bridge\AmazonSns\AmazonSnsTransport;
+use Symfony\Component\Notifier\Exception\InvalidArgumentException;
 use Symfony\Component\Notifier\Message\ChatMessage;
 use Symfony\Component\Notifier\Message\MessageInterface;
 use Symfony\Component\Notifier\Message\MessageOptionsInterface;
@@ -44,6 +45,16 @@ class AmazonSnsTransportTest extends TransportTestCase
     {
         yield [$this->createMock(MessageInterface::class)];
         yield [new ChatMessage('hello', $this->createMock(MessageOptionsInterface::class))];
+    }
+
+    public function testSmsMessageWithFrom()
+    {
+        $transport = $this->createTransport();
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The "Symfony\Component\Notifier\Bridge\AmazonSns\AmazonSnsTransport" transport does not support "from" in "Symfony\Component\Notifier\Message\SmsMessage".');
+
+        $transport->send(new SmsMessage('0600000000', 'test', 'foo'));
     }
 
     public function testSmsMessageOptions()