+++ src/Symfony/Component/Notifier/Bridge/Clickatell/ClickatellTransport.php
@@ -61,6 +61,8 @@ final class ClickatellTransport extends AbstractTransport
 
         $endpoint = sprintf('https://%s/rest/message', $this->getEndpoint());
 
+        $from = $message->getFrom() ?: $this->from;
+
         $response = $this->client->request('POST', $endpoint, [
             'headers' => [
                 'Accept' => 'application/json',
@@ -69,7 +71,7 @@ final class ClickatellTransport extends AbstractTransport
                 'X-Version' => 1,
             ],
             'json' => [
-                'from' => $this->from ?? '',
+                'from' => $from ?? '',
                 'to' => [$message->getPhone()],
                 'text' => $message->getSubject(),
             ],