+++ src/Symfony/Component/Notifier/Bridge/FortySixElks/FortySixElksTransport.php
@@ -57,10 +57,12 @@ final class FortySixElksTransport extends AbstractTransport
             throw new UnsupportedMessageTypeException(__CLASS__, SmsMessage::class, $message);
         }
 
+        $from = $message->getFrom() ?: $this->from;
+
         $endpoint = sprintf('https://%s/a1/sms', self::HOST);
         $response = $this->client->request('POST', $endpoint, [
             'body' => [
-                'from' => $this->from,
+                'from' => $from,
                 'to' => $message->getPhone(),
                 'message' => $message->getSubject(),
             ],