+++ src/Symfony/Component/Notifier/Bridge/FreeMobile/FreeMobileTransport.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Notifier\Bridge\FreeMobile;
 
+use Symfony\Component\Notifier\Exception\InvalidArgumentException;
 use Symfony\Component\Notifier\Exception\TransportException;
 use Symfony\Component\Notifier\Exception\UnsupportedMessageTypeException;
 use Symfony\Component\Notifier\Message\MessageInterface;
@@ -55,6 +56,11 @@ final class FreeMobileTransport extends AbstractTransport
     {
         if (!$this->supports($message)) {
             throw new UnsupportedMessageTypeException(__CLASS__, SmsMessage::class, $message);
+        }
+
+        /** @var SmsMessage $message */
+        if ('' !== $message->getFrom()) {
+            throw new InvalidArgumentException(sprintf('The "%s" transport does not support "from" in "%s".', __CLASS__, SmsMessage::class));
         }
 
         $endpoint = sprintf('https://%s', $this->getEndpoint());