+++ src/Symfony/Component/Notifier/Bridge/Mobyt/MobytTransport.php
@@ -74,7 +74,11 @@ final class MobytTransport extends AbstractTransport
         $options['message'] ??= $message->getSubject();
         $options['recipient'] = [$message->getPhone()];
 
-        $options['sender'] ??= $this->from;
+        if ('' !== $message->getFrom()) {
+            $options['sender'] = $message->getFrom();
+        } else {
+            $options['sender'] ??= $this->from;
+        }
 
         $response = $this->client->request('POST', 'https://'.$this->getEndpoint().'/API/v1.0/REST/sms', [
             'headers' => [