+++ src/Symfony/Component/Notifier/Bridge/Octopush/OctopushTransport.php
@@ -59,6 +59,8 @@ final class OctopushTransport extends AbstractTransport
             throw new UnsupportedMessageTypeException(__CLASS__, SmsMessage::class, $message);
         }
 
+        $from = $message->getFrom() ?: $this->from;
+
         $endpoint = sprintf('https://%s/api/sms/json', $this->getEndpoint());
 
         $response = $this->client->request('POST', $endpoint, [
@@ -70,7 +72,7 @@ final class OctopushTransport extends AbstractTransport
                 'api_key' => $this->apiKey,
                 'sms_text' => $message->getSubject(),
                 'sms_recipients' => $message->getPhone(),
-                'sms_sender' => $this->from,
+                'sms_sender' => $from,
                 'sms_type' => $this->type,
             ],
         ]);