+++ src/Symfony/Component/Notifier/Bridge/Sinch/SinchTransport.php
@@ -57,11 +57,13 @@ final class SinchTransport extends AbstractTransport
             throw new UnsupportedMessageTypeException(__CLASS__, SmsMessage::class, $message);
         }
 
+        $from = $message->getFrom() ?: $this->from;
+
         $endpoint = sprintf('https://%s/xms/v1/%s/batches', $this->getEndpoint(), $this->accountSid);
         $response = $this->client->request('POST', $endpoint, [
             'auth_bearer' => $this->authToken,
             'json' => [
-                'from' => $this->from,
+                'from' => $from,
                 'to' => [$message->getPhone()],
                 'body' => $message->getSubject(),
             ],