+++ src/Symfony/Component/Notifier/Bridge/Smsc/SmscTransport.php
@@ -59,10 +59,12 @@ final class SmscTransport extends AbstractTransport
             throw new UnsupportedMessageTypeException(__CLASS__, SmsMessage::class, $message);
         }
 
+        $from = $message->getFrom() ?: $this->from;
+
         $body = [
             'login' => $this->login,
             'psw' => $this->password,
-            'sender' => $this->from,
+            'sender' => $from,
             'phones' => $message->getPhone(),
             'mes' => $message->getSubject(),
             'fmt' => 3, // response as JSON